% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Gerador do dashboard em shiny}
\usage{
app()
}
\description{
Esta função gera o dashboard em shiny que foi desenvolvido para visualizar os dados das partidas do Brasileirão de pontos corridos
}
