% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{coresBR}
\alias{coresBR}
\title{Cores dos Times Brasileiros}
\format{
Vetor com 47 elementos, todos do tipo string
}
\usage{
coresBR()
}
\description{
Hexcode da cor principal do brasão de cada time que passou pela Série A do Brasileirão de 2003 à 2023
}
